requirements.txt
matplotlib
pandas
numpy
Pillow
setuptools

-e .
